clean up app deployment scripts

+++ apps/alacritty/deploy.sh
@@ -8,11 +8,22 @@ if ! [ -d "src" ]; then
     git clone $REPO src
 fi
 cd src
-git pull
 
 # build and save to applications
-make app
-cp -r target/release/osx/Alacritty.app $HOME/Applications
+if [ "$1" == "--force" ]; then
+    make app
+    cp -r target/release/osx/Alacritty.app $HOME/Applications
+else
+    git fetch
+    ALACRITTY_VERSION=$(git tag | tail -n 1)
+    CURRENT_COMMIT=$(git rev-parse @)
+    VERSION_COMMIT=$(git rev-list -n 1 $ALACRITTY_VERSION)
+    if [ $LOCAL_COMMIT != $REMOTE_COMMIT ]; then
+        git checkout $ALACRITTY_VERSION
+        make app
+        cp -r target/release/osx/Alacritty.app $HOME/Applications
+    fi
+fi
 
 if ! [ -d "$HOME/.config/alacritty" ]; then
     echo "creating alacritty directory in ~/.config"
